api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
SQLAlchemy==2.0.34
asyncpg==0.29.0
psycopg[binary]==3.2.1
python-multipart==0.0.9
azure-storage-blob==12.20.0
aio-pika==9.4.3
httpx==0.27.2
structlog==24.1.0
